mcps/hormozi-mcp/requirements.txt
mcp[cli]>=1.23.0
requests
openai
httpx
starlette
uvicorn
python-dotenv